requirements.txt
# Used for rendering final HTML from template
Jinja2==3.1.2

# Used for parsing Markdown content
markdown==3.3.3

# Used for parsing YAML frontmatter
python-frontmatter==1.0.0
